chore: revert composer footer to part of body
