+++ apps/meteor/client/views/room/RoomSkeleton.tsx
@@ -3,10 +3,18 @@ import React from 'react';
 
 import MessageListSkeleton from '../../components/message/list/MessageListSkeleton';
 import HeaderSkeleton from './Header/HeaderSkeleton';
-import ComposerSkeleton from './footer/RoomFooterSkeleton';
+import RoomComposerSkeleton from './composer/RoomComposer/RoomComposerSkeleton';
 import RoomLayout from './layout/RoomLayout';
 
 const RoomSkeleton = (): ReactElement => (
-	<RoomLayout header={<HeaderSkeleton />} body={<MessageListSkeleton />} footer={<ComposerSkeleton />} />
+	<RoomLayout
+		header={<HeaderSkeleton />}
+		body={
+			<>
+				<MessageListSkeleton />
+				<RoomComposerSkeleton />
+			</>
+		}
+	/>
 );
 export default RoomSkeleton;
