+++ apps/meteor/client/views/room/composer/RoomComposer/RoomComposerSkeleton.tsx
@@ -1,0 +1,12 @@
+import type { FC } from 'react';
+import React, { memo } from 'react';
+
+import ComposerSkeleton from '../ComposerSkeleton';
+import RoomComposer from './RoomComposer';
+
+const RoomComposerSkeleton: FC = () => (
+	<RoomComposer>
+		<ComposerSkeleton />
+	</RoomComposer>
+);
+export default memo(RoomComposerSkeleton);
